requirements.txt
flask
flasgger
tensorflow
numpy
python-dotenv
